feat: LunchCampusSelectBox 분리
