+++ src/components/Lunch/index.ts
@@ -1,3 +1,4 @@
 export * from './LunchTabs';
 export * from './LunchCard';
 export * from './LunchIntroduction';
+export * from './LunchCampusSelectBox';
