+++ src/pages/lunch/index.tsx
@@ -3,15 +3,19 @@ import { useRouter } from 'next/router';
 import { css } from '@emotion/react';
 import React from 'react';
 
-import { DefaultFullPageLoader, SelectBox } from '~/components/Common';
+import { DefaultFullPageLoader } from '~/components/Common';
 import { LunchLayout } from '~/components/Layout';
-import { LunchTabs, LunchIntroduction } from '~/components/Lunch';
+import {
+  LunchTabs,
+  LunchIntroduction,
+  LunchCampusSelectBox,
+} from '~/components/Lunch';
 import LunchMenus from '~/components/Lunch/LunchMenus';
 import NavigationGroup from '~/components/NavigationGroup';
 import { LunchDateSpecifier } from '~/services/lunch/utils';
 import { useCampuses } from '~/services/meta';
 
-const validateDateSpecifier = (date: string): date is LunchDateSpecifier => {
+const validateDateSpecifier = (date?: string): date is LunchDateSpecifier => {
   return Object.values(LunchDateSpecifier).includes(date as LunchDateSpecifier);
 };
 
@@ -25,16 +29,17 @@ const Lunch = () => {
   const query = router.query as Partial<QueryString>;
 
   const { data: campuses } = useCampuses();
-  const { campus, date } = query;
+  const { campus, date: dateSpecifier } = query;
 
   const isValidCampus = campus && campuses.includes(campus);
-  const isValidDateSpecifier = date && validateDateSpecifier(date);
+  const isValidDateSpecifier =
+    dateSpecifier && validateDateSpecifier(dateSpecifier);
   const isValidQueryParams = isValidCampus && isValidDateSpecifier;
 
   if (!isValidQueryParams) {
     const safeCampus = isValidCampus ? campus : campuses[0];
     const safeDateSpecifier = isValidDateSpecifier
-      ? date
+      ? dateSpecifier
       : LunchDateSpecifier.TODAY;
 
     router.push({
@@ -63,13 +68,18 @@ const Lunch = () => {
     <LunchLayout>
       <NavigationGroup />
       <LunchIntroduction />
-      <CampusSelectBox
+      <LunchCampusSelectBox
+        css={lunchCampusSelectBoxCss}
         selectedCampus={campus}
         campuses={campuses}
         onCampusChange={onCampusChange}
       />
       <LunchTabs />
-      <LunchMenus css={lunchMenusCss} campus={campus} dateSpecifier={date} />
+      <LunchMenus
+        css={lunchMenusCss}
+        campus={campus}
+        dateSpecifier={dateSpecifier}
+      />
     </LunchLayout>
   );
 };
@@ -80,40 +90,6 @@ const lunchPageZIndex = {
 };
 
 const lunchMenusCss = css({ zIndex: lunchPageZIndex.lunchMenu });
+const lunchCampusSelectBoxCss = css({ zIndex: lunchPageZIndex.selectBox });
 
 export default Lunch;
-
-interface CampusSelectBoxProps {
-  className?: string;
-  selectedCampus: string;
-  campuses: string[];
-  onCampusChange: (value: string) => void;
-}
-
-const CampusSelectBox = (props: CampusSelectBoxProps) => {
-  const { selectedCampus, campuses, onCampusChange, className } = props;
-
-  return (
-    <div css={campusSelectBoxSelfCss} className={className}>
-      <SelectBox
-        items={campuses}
-        size="lg"
-        placeholder="캠퍼스 선택"
-        triggerTextAlign="center"
-        value={selectedCampus}
-        onValueChange={onCampusChange}
-      />
-    </div>
-  );
-};
-
-const campusSelectBoxSelfCss = css({
-  width: '100%',
-  zIndex: lunchPageZIndex.selectBox,
-});
-
-export async function getServerSideProps() {
-  return {
-    props: { data: 1 },
-  };
-}
